Revert "resolved conflicts"
This reverts commit e86f1d2a4a4d81faf12080457d6249fbd49808f9.

+++ Microsoft.Toolkit.Uwp.Notifications/Toasts/Compat/ToastNotificationManagerCompat.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -27,7 +27,6 @@ namespace Microsoft.Toolkit.Uwp.Notifications
 #if WIN32
         private const string TOAST_ACTIVATED_LAUNCH_ARG = "-ToastActivated";
         private const string REG_HAS_SENT_NOTIFICATION = "HasSentNotification";
-        private const string REG_HAS_7_0_1_FIX = "Has7.0.1Fix";
         internal const string DEFAULT_GROUP = "toolkitGroupNull";
 
         private const int CLASS_E_NOAGGREGATION = -2147221232;
@@ -176,23 +175,6 @@ namespace Microsoft.Toolkit.Uwp.Notifications
 
                     // Additionally, we need to read whether they've sent a notification before
                     _hasSentNotification = rootKey.GetValue(REG_HAS_SENT_NOTIFICATION) != null;
-
-                    // And read if we've already applied the 7_0_1 fix
-                    bool has7_0_1fix = rootKey.GetValue(REG_HAS_7_0_1_FIX) != null;
-
-                    // If it doesn't have the fix yet
-                    if (!has7_0_1fix)
-                    {
-                        // If the AUMID changed
-                        if (win32AppInfo.Pre7_0_1Aumid != null)
-                        {
-                            // Uninstall the old AUMID
-                            CleanUpOldAumid(win32AppInfo.Pre7_0_1Aumid);
-                        }
-
-                        // Set that it has the fix so we don't try uninstalling again in the future
-                        rootKey.SetValue(REG_HAS_7_0_1_FIX, 1);
-                    }
                 }
 
                 rootKey.SetValue("CustomActivator", string.Format("{{{0}}}", activatorType.GUID));
@@ -201,12 +183,7 @@ namespace Microsoft.Toolkit.Uwp.Notifications
 
         private static string GetRegistrySubKey()
         {
-            return GetRegistrySubKey(_win32Aumid);
-        }
-
-        private static string GetRegistrySubKey(string win32Aumid)
-        {
-            return @"Software\Classes\AppUserModelId\" + win32Aumid;
+            return @"Software\Classes\AppUserModelId\" + _win32Aumid;
         }
 
         private static Type CreateActivatorType()
@@ -466,7 +443,7 @@ namespace Microsoft.Toolkit.Uwp.Notifications
                 {
                     // Remove all scheduled notifications (do this first before clearing current notifications)
                     var notifier = CreateToastNotifier();
-                    foreach (var scheduled in notifier.GetScheduledToastNotifications())
+                    foreach (var scheduled in CreateToastNotifier().GetScheduledToastNotifications())
                     {
                         try
                         {
@@ -539,70 +516,20 @@ namespace Microsoft.Toolkit.Uwp.Notifications
             {
             }
 
-            try
+            if (!DesktopBridgeHelpers.HasIdentity() && _win32Aumid != null)
             {
-                // Delete any of the app files
-                var appDataFolderPath = Win32AppInfo.GetAppDataFolderPath(_win32Aumid);
-                if (Directory.Exists(appDataFolderPath))
+                try
                 {
-                    Directory.Delete(appDataFolderPath, recursive: true);
-                }
-            }
-            catch
-            {
-            }
-        }
-
-        private static void CleanUpOldAumid(string oldAumid)
-        {
-            try
-            {
-                // Remove all scheduled notifications (do this first before clearing current notifications)
-                var notifier = ToastNotificationManager.CreateToastNotifier(oldAumid);
-                foreach (var scheduled in notifier.GetScheduledToastNotifications())
-                {
-                    try
+                    // Delete any of the app files
+                    var appDataFolderPath = Win32AppInfo.GetAppDataFolderPath(_win32Aumid);
+                    if (Directory.Exists(appDataFolderPath))
                     {
-                        notifier.RemoveFromSchedule(scheduled);
-                    }
-                    catch
-                    {
+                        Directory.Delete(appDataFolderPath, recursive: true);
                     }
                 }
-            }
-            catch
-            {
-            }
-
-            try
-            {
-                // Clear all current notifications
-                ToastNotificationManager.History.Clear(oldAumid);
-            }
-            catch
-            {
-            }
-
-            try
-            {
-                // Remove registry key
-                Registry.CurrentUser.DeleteSubKey(GetRegistrySubKey(oldAumid));
-            }
-            catch
-            {
-            }
-
-            try
-            {
-                // Delete any of the app files
-                var appDataFolderPath = Win32AppInfo.GetAppDataFolderPath(oldAumid);
-                if (Directory.Exists(appDataFolderPath))
+                catch
                 {
-                    Directory.Delete(appDataFolderPath, recursive: true);
                 }
-            }
-            catch
-            {
             }
         }
 #endif
